requirements.txt
fastapi
uvicorn
numpy
opencv-python
ultralytics
pyttsx3
python-multipart
pillow